worked on navbar
