+++ src/scripts/events/domEvents.js
@@ -1,5 +1,5 @@
 const domEvents = () => {
 
-}
+};
 
 export default domEvents;
