+++ src/scripts/events/navigationEvents.js
@@ -1,0 +1,28 @@
+/*
+import signOut from '../helpers/auth/signOut';
+
+// navigation events
+const navigationEvents = () => {
+  // LOGOUT BUTTON
+  document.querySelector('#logout-button')
+    .addEventListener('click', signOut);
+
+  // VIEW ORDER
+  document.querySelector('#view-order').addEventListener('click', () => {
+    viewOrder().then((ordersArray) => showOrder(ordersArray));
+  });
+
+  // CREATE AN ORDER
+  document.querySelector('#create-order').addEventListener('click', () => {
+    createOrder().then((ordersArray) => showOrder(ordersArray));
+  });
+
+  // SEARCH
+  document.querySelector('#search').addEventListener('keyup', (e) => {
+    const searchValue = document.querySelector('#search').value.toLowerCase();
+    console.warn(searchValue);
+  });
+};
+
+export default navigationEvents;
+*/
